MOD: prueba

+++ vtex/orders_api/ORDERS/descomposicion_orders/item.py
@@ -224,7 +224,7 @@ def run():
         "mode": "NULLABLE"
     },{
         "name": "sellingPrice",
-        "type": "FLOAT",
+        "type": "INTEGER",
         "mode": "NULLABLE"
     },{
         "name": "name",
